+++ apps/web/components/OfferPageContent/types.ts
@@ -1,0 +1,6 @@
+import { Offer } from '@plentymarkets/shop-api';
+
+export type OfferPageContentProps = {
+  offer: Offer;
+  offerLoading: boolean;
+};
